incomeexpenseapi/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docopt==0.4.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
mailchimp==2.0.10
mailchimp-marketing==3.0.80
mailchimp-transactional==1.0.56
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
